refactor(annotations): migrate annotations to wow-api and deprecate wow-spring stereotypes
- Move @EventProcessor, @ProjectionProcessor, and @StatelessSaga annotations to wow-api
- Add @Component annotation to wow-api annotations for automatic component scanning
- Deprecate corresponding annotations in wow-spring, marking them for removal
- Update example usage to use new wow-api annotations
- Remove redundant spring.stereotype imports in example files
